refactor: Extract graph utils to separate module
