+++ libs/rdf/src/main/java/no/sikt/nva/nvi/rdf/NvaOntology.java
@@ -1,0 +1,9 @@
+package no.sikt.nva.nvi.rdf;
+
+public final class NvaOntology {
+
+  public static final String PART_OF = "https://nva.sikt.no/ontology/publication#partOf";
+  public static final String HAS_PART = "https://nva.sikt.no/ontology/publication#hasPart";
+
+  private NvaOntology() {}
+}
