+++ libs/rdf/src/test/java/no/sikt/nva/nvi/rdf/JsonLdModelsTest.java
@@ -1,6 +1,6 @@
-package no.sikt.nva.nvi.viewingscope.utils;
+package no.sikt.nva.nvi.rdf;
 
-import static no.sikt.nva.nvi.viewingscope.utils.GraphUtils.createModel;
+import static no.sikt.nva.nvi.rdf.JsonLdModels.createModel;
 import static no.unit.nva.commons.json.JsonUtils.dtoObjectMapper;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
@@ -8,7 +8,7 @@ import java.nio.file.Path;
 import nva.commons.core.ioutils.IoUtils;
 import org.junit.jupiter.api.Test;
 
-class GraphUtilsTest {
+class JsonLdModelsTest {
 
   @Test
   void shouldNotThrowDuringCreateModel() {
