+++ libs/rdf/src/test/resources/candidate.json
